Gw iter gpu

+++ pyscf/nao/chi0_matvec.py
@@ -17,7 +17,7 @@ class chi0_matvec(mf):
   def __init__(self, **kw):
     from pyscf.nao.m_fermi_dirac import fermi_dirac_occupations
 
-    self.dtype = kw['dtype'] if 'dtype' in kw else np.float32
+    self.dtype = kw['dtype'] if 'dtype' in kw else np.float64
     for x in ['dtype']: kw.pop(x, None)
     mf.__init__(self, dtype=self.dtype, **kw)
 
@@ -85,7 +85,9 @@ class chi0_matvec(mf):
       
     pb = self.pb
     self.moms0,self.moms1 = pb.comp_moments(dtype=self.dtype)
-    self.td_GPU = tddft_iter_gpu_c(GPU, self.mo_coeff[0,0,:,:,0], self.ksn2f, self.ksn2e, self.norbs, self.nfermi, self.nprod, self.vstart)
+    self.td_GPU = tddft_iter_gpu_c(GPU, self.mo_coeff[0,0,:,:,0], self.ksn2f, 
+                                   self.ksn2e, self.norbs, self.nfermi, self.nprod,
+                                   self.vstart)
 
   def apply_rf0(self, sp2v, comega=1j*0.0):
     """ This applies the non-interacting response function to a vector (a set of vectors?) """
